backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
requests>=2.32.0
python-multipart>=0.0.12